Add error notification interface & implementation
This commit adds a generic interface for all the error notification
types, which could then be provided to the error handler, as appropriate
to the query type.

The content error layout has also been updated to contain an action
button, and to set it's visibility to gone by default so that it's not
unnecessarily set up as part of the view hierarchy layout.

The AlertDialogFragment has also been updated to include a new static
constructor that takes string resource IDs instead of the actual
resolved strings, in order to not have an unneeded dependency on the
Context in it's notification implementation.